EmailProcessorGPT/requirements.txt
flask
python-dotenv
spacy
requests
openai==0.28
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-storage
SQLAlchemy
utils
simplejson
APScheduler
boolean
pymongo
retrying
storage
typing 

##python -m spacy download en_core_web_sm